SpreadsheetEngineHateosHandler sub-class naming refactor
- Only renamed classes, no logic changes.
- Introduced abstract SpreadsheetEngineHateosHandlerSpreadsheetDelta
